+++ src/sim/simul_solve.cc
@@ -231,7 +231,7 @@ void Simul::solve_auto()
     
     // solve the system, recording time:
     //double cpu = TimeDate::milliseconds();
-    size_t cnt = sMeca.solve();
+    unsigned cnt = sMeca.solve();
     //float cpu = TimeDate::milliseconds() - cpu;
     // use Meca::cycles_ that only includes preconditionning parts!
     float cpu = sMeca.cycles_;
